[fontc] Move ChangeDetector into change_detector.rs
